back/requirements.txt
Django==5.0.3
djangorestframework==3.14
django-cors-headers==4.3.0
python-dotenv==1.0.1
psycopg[binary]
django_extensions==3.2.3
Pillow==10.4.0
markdown2[all]
channels==4.1.0
daphne==4.1.2